refactor(runtimes): add torchtune.go
